+++ bin/dev
@@ -1,0 +1,2 @@
+require('babel-register')
+require('./../server/server')
