+++ bin/production
@@ -1,0 +1,1 @@
+require('./../dist/server')
