+++ package.json
@@ -8,8 +8,11 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server/server",
-    "dev": "nodemon server/server"
+    "start": "node bin/dev",
+    "dev": "nodemon bin/dev",
+    "clean": "rm -rf dist",
+    "build": "npm run clean && mkdir dist && babel server -s -d dist",
+    "production": "npm run build && node bin/production"
   },
   "repository": {
     "type": "git",
@@ -27,6 +30,10 @@
     "uuid": "^3.3.2"
   },
   "devDependencies": {
+    "@babel/register": "^7.4.0",
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-register": "^6.26.0",
     "nodemon": "^1.18.10"
   }
 }
