+++ server/models/Users.js
@@ -32,4 +32,4 @@ const users = [
 ];
 
 
-module.exports = users;
+export default users;
